Normalize the usage of long statistics

+++ app/src/main/java/es/danirod/rectball/screens/GameOverScreen.java
@@ -54,8 +54,8 @@ public class GameOverScreen extends AbstractScreen {
         // Set up the label data.
         String lastScore = Integer.toString(game.getState().getScore());
         while (lastScore.length() < 4) lastScore = "0" + lastScore;
-        String aliveTime = Integer.toString(Math.round(game.getState().getElapsedTime()));
-        String highScore = Integer.toString(game.getSettings().getPreferences().getInt(SettingsManager.TAG_HIGH_SCORE, 0));
+        String aliveTime = Long.toString(Math.round(game.getState().getElapsedTime()));
+        String highScore = Long.toString(game.getSettings().getPreferences().getLong(SettingsManager.TAG_HIGH_SCORE, 0L));
 
         // Scores table
         Table scoresTable = new Table();
